made student endpoint and deleted things
