+++ server/api/student/student.socket.js
@@ -4,21 +4,21 @@
 
 'use strict';
 
-var thing = require('./thing.model');
+var Student = require('./student.model');
 
 exports.register = function(socket) {
-  thing.schema.post('save', function (doc) {
+  Student.schema.post('save', function (doc) {
     onSave(socket, doc);
   });
-  thing.schema.post('remove', function (doc) {
+  Student.schema.post('remove', function (doc) {
     onRemove(socket, doc);
   });
 }
 
 function onSave(socket, doc, cb) {
-  socket.emit('thing:save', doc);
+  socket.emit('student:save', doc);
 }
 
 function onRemove(socket, doc, cb) {
-  socket.emit('thing:remove', doc);
+  socket.emit('student:remove', doc);
 }
