+++ server/api/student/student.spec.js
@@ -4,11 +4,11 @@ var should = require('should');
 var app = require('../../app');
 var request = require('supertest');
 
-describe('GET /api/things', function() {
+describe('GET /api/students', function() {
 
   it('should respond with JSON array', function(done) {
     request(app)
-      .get('/api/things')
+      .get('/api/students')
       .expect(200)
       .expect('Content-Type', /json/)
       .end(function(err, res) {
